# lib/api-zod/src/generated/types/activeSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */

export interface ActiveSession {
  sessionId: string;
  currentPage: string;
  startedAt: string;
  /** @nullable */
  country?: string | null;
}