# lib/api-zod/src/generated/types/activeUsers.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */
import type { ActiveSession } from './activeSession';

export interface ActiveUsers {
  count: number;
  sessions: ActiveSession[];
}