# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  clientPhone?: string | null;
  designType: string;
  description: string;
  /** @nullable */
  references?: string | null;
  status: OrderStatus;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  deadline?: string | null;
  createdAt: string;
}