# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  clientName: string;
  clientEmail: string;
  clientPhone?: string;
  designType: string;
  description: string;
  references?: string;
  budget?: string;
  deadline?: string;
}