# lib/api-zod/src/generated/types/portfolioWorkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioWorkInput {
  title: string;
  titleAr: string;
  clientName: string;
  category: string;
  imageUrl: string;
  description?: string;
  featured?: boolean;
  designer?: string;
}