# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  totalVisitors: number;
  totalWorks: number;
  totalOrders: number;
  pendingOrders: number;
  totalUsers: number;
  monthlyVisitors: number;
}