# lib/api-zod/src/generated/types/visitorLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shadj API - شدج
 * OpenAPI spec version: 0.1.0
 */

export interface VisitorLog {
  id: string;
  sessionId: string;
  journey: string[];
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  device?: string | null;
}